pyproject.toml
[tool.poetry]
name = "backup-builder"
version = "0.1.0"
description = ""
authors = ["Danila Kislitsyn <dangost16@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
